[STRUCT] Organize_code_functions_into_other_files
