+++ resources/js/Components/Profile/Who'sOnline.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function OnlinePerson({ name, imageUrl }) {
+    return (
+        <img
+          loading="lazy"
+          src={imageUrl}
+          alt={name}
+          className="shrink-0 w-11 aspect-[0.95]"
+        />
+    );
+}
+
+export default OnlinePerson;
