+++ resources/js/Components/Profile/index.js
@@ -1,0 +1,4 @@
+export { default as ProfileHeader } from "./Header";
+export { default as ProfileNav } from "./Nav";
+export { default as Popup } from "./BannerPopup";
+export { default as OnlinePerson } from "./Who'sOnline";
