+++ resources/js/Components/ProfileTabbar/index.js
@@ -1,0 +1,4 @@
+export { default as ProfileBio } from "../ProfileTabbar/Bio";
+export { default as ProfileGallery } from "../ProfileTabbar/Gallery";
+export { default as ProfileIcons } from "./BioIcons";
+export { SearchInput, SearchButton } from "./Files";
