+++ resources/js/Components/SideCalendar/index.js
@@ -1,0 +1,1 @@
+export { default as FeaturedEvent } from "./FeaturedEvents";
