+++ resources/js/Pages/Profile.jsx
@@ -1,326 +1,8 @@
 import * as React from "react";
+import { ProfileHeader, ProfileNav, Popup, OnlinePerson } from "../Components/Profile";
+import { ProfileBio, ProfileGallery, ProfileIcons, SearchInput, SearchButton } from "../Components/ProfileTabbar"
+import { FeaturedEvent } from "../Components/SideCalendar";
 
-function ProfileImage({ src, alt, className }) {
-  return (
-    <div className={`flex overflow-hidden relative z-10 flex-col items-end px-16 pt-20 pb-3.5 mt-44 mb-0 w-44 max-w-full aspect-square max-md:px-5 max-md:mt-10 max-md:mb-2.5 ${className}`}>
-      <img src={src} alt={alt} className="object-cover absolute inset-0 size-full" />
-      <div className="relative shrink-0 mt-16 w-5 h-5 bg-green-500 rounded-full border-2 border-white border-solid stroke-[2px] max-md:mt-10 left-10" />
-    </div>
-  );
-}
-
-function ProfileHeader({ backgroundImage, profileImage, name, status, onEditBanner }) {
-  const handleEditBanner = (e) => {
-    e.stopPropagation(); // Prevents the click event from propagating to parent elements
-    onEditBanner(); // Calls the onEditBanner function when the background image is clicked
-  };
-  return (
-    <header
-      className="flex overflow-hidden relative z-10 flex-col items-start px-7 pt-20 -mt-14 w-half min-h-[270px] max-md:px-5 max-md:max-w-half"
-      onClick={handleEditBanner} // Added onClick handler to trigger onEditBanner
-    >
-      <img src={backgroundImage} alt="" className="object-cover absolute inset-0 size-half" />
-      <ProfileImage src={profileImage} alt={`${name}'s profile picture`} />
-      <div className="flex flex-col self-center px-5 -mt-10 -ml-96">
-        <h1 className="self-end text-3xl font-extrabold text-neutral-800">{name}</h1>
-        <div className="mt-2 text-xs font-semibold text-neutral-800 text-opacity-50">{status}</div>
-      </div>
-    </header>
-  );
-}
-
-function ProfileNav({ activeTab, setActiveTab }) {
-  return (
-    <nav className="flex gap-5 justify-between mt-14 ml-10 max-w-full text-sm font-semibold text-center whitespace-nowrap text-neutral-800 text-opacity-30 w-[423px] max-md:flex-wrap max-md:mt-10">
-      <a href="#activities" className={`text-stone-300 ${activeTab === "activities" ? "font-bold text-blue-500" : ""}`} onClick={() => setActiveTab("activities")}>Activities</a>
-      <a href="#bio" className={`text-base font-bold ${activeTab === "bio" ? "text-blue-500" : ""}`} onClick={() => setActiveTab("bio")}>Bio</a>
-      <a href="#gallery" className={`${activeTab === "gallery" ? "font-bold text-blue-500" : ""}`} onClick={() => setActiveTab("gallery")}>Gallery</a>
-      <a href="#files" className={`${activeTab === "files" ? "font-bold text-blue-500" : ""}`} onClick={() => setActiveTab("files")}>Files</a>
-    </nav>
-  );
-}
-
-function ProfileBio({ email, department, position, grade, location, phone }) {
-  return (
-    <div className="flex-auto my-auto max-md:max-w-full">
-      <div className="flex gap-5 max-md:flex-col max-md:gap-0">
-        <div className="flex flex-col w-[36%] max-md:ml-0 max-md:w-full">
-          <div className="flex flex-col grow text-base font-semibold capitalize text-neutral-800 max-md:mt-10">
-            <div>e-mail</div>
-            <div className="mt-6">department</div>
-            <div className="mt-8">position</div>
-            <div className="mt-7">grade</div>
-            <div className="mt-7">location</div>
-            <div className="mt-7">mob/Ext number</div>
-          </div>
-        </div>
-        <div className="flex flex-col ml-5 w-[64%] max-md:ml-0 max-md:w-full">
-          <div className="flex flex-col grow text-sm text-neutral-800 text-opacity-80 max-md:mt-10">
-            <div>{email}</div>
-            <div className="mt-8">{department}</div>
-            <div className="mt-8">{position}</div>
-            <div className="mt-8">{grade}</div>
-            <div className="mt-8">{location}</div>
-            <div className="mt-8">{phone}</div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function PhotoItem({ src, alt }) {
-  return (
-    <div className="flex flex-col w-3/12 max-md:ml-0 max-md:w-full">
-      <img
-        loading="lazy"
-        src={src}
-        alt={alt}
-        className="grow shrink-0 mt-8 max-w-full aspect-[1.19] w-[173px] max-md:mt-10"
-      />
-    </div>
-  );
-}
-
-function VideoItem({ src, alt }) {
-  return (
-    <div className="flex flex-col w-3/12 max-md:ml-0 max-md:w-full">
-      <img
-        loading="lazy"
-        src={src}
-        alt={alt}
-        className="grow shrink-0 mt-8 max-w-full aspect-[1.19] w-[173px] max-md:mt-10"
-      />
-    </div>
-  );
-}
-
-function ProfileGallery({ photoData, videoData }) {
-  return (
-    <div className="flex flex-col max-w-[800px]">
-      <section className="px-6 py-6 w-full bg-white rounded-md shadow-sm max-md:px-5 max-md:max-w-full">
-        <div className="flex gap-5 max-md:flex-col max-md:gap-0">
-          <div className="flex flex-col w-3/12 max-md:ml-0 max-md:w-full">
-            <div className="flex flex-col grow text-2xl font-bold whitespace-nowrap text-neutral-800 max-md:mt-10">
-              <h2>Photo</h2>
-              <img
-                loading="lazy"
-                src="https://cdn.builder.io/api/v1/image/assets/TEMP/b964b38ab4577424894ab89afca98d210e5a3ab6ee6f4091065c1b0e53df2748?apiKey=23ce5a6ac4d345ebaa82bd6c33505deb&"
-                alt="Main Photo"
-                className="self-center mt-3.5 aspect-[1.19] w-[172px]"
-              />
-            </div>
-          </div>
-          {photoData.map((photo, index) => (
-            <PhotoItem key={index} src={photo.src} alt={photo.alt} />
-          ))}
-        </div>
-      </section>
-      <section className="px-6 py-6 w-full bg-white rounded-md shadow-sm max-md:px-5 max-md:max-w-full">
-        <div className="flex gap-5 max-md:flex-col max-md:gap-0">
-          <div className="flex flex-col w-3/12 max-md:ml-0 max-md:w-full">
-            <div className="flex flex-col grow text-2xl font-bold whitespace-nowrap text-neutral-800 max-md:mt-10">
-              <h2>Videos</h2>
-              <img
-                loading="lazy"
-                src="https://cdn.builder.io/api/v1/image/assets/TEMP/cdbbaca5c344dcb79e33b324a787c8c2119e2929aebc1bda0bf551ae62ef74fc?apiKey=23ce5a6ac4d345ebaa82bd6c33505deb&"
-                alt="Video 1"
-                className="self-center mt-3.5 aspect-[1.92] w-[172px]"
-              />
-            </div>
-          </div>
-          {videoData.map((video, index) => (
-            <VideoItem key={index} src={video.src} alt={video.alt} />
-          ))}
-        </div>
-      </section>
-    </div>
-  );
-}
-
-function SearchInput() {
-  return (
-    <div className="flex min-w-72 gap-2 px-5 py-1.5 text-md bg-white rounded-full border border-solid border-neutral-200 text-neutral-800 text-opacity-50 mt-8">
-      <img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/e9a74adec558c689d2c2036311d5dad5a1cc1d44aea0cf5f88c1bd1bafeea3ce?apiKey=285d536833cc4168a8fbec258311d77b&"
-        alt=""
-        className="shrink-0 w-6 aspect-square"
-      />
-      <input
-        type="text"
-        className="flex-auto bg-transparent border-none outline-none text-md text-neutral-800"
-        placeholder="Search"
-      />
-    </div>
-  );
-}
-
-function SearchButton() {
-  return (
-    <button className="justify-center px-5 py-1.5 my-auto text-s font-bold text-center text-white bg-blue-500 rounded-3xl mt-10">
-      Search
-    </button>
-  );
-}
-
-function ProfileIcons({ icon1, icon2 }) {
-  const [isPopupOpen, setIsPopupOpen] = React.useState(false);
-  const [qrCodeLink, setQrCodeLink] = React.useState('');
-
-  const openPopup = () => {
-    setIsPopupOpen(true);
-    console.log("bukak");
-  };
-
-  const closePopup = () => {
-    setIsPopupOpen(false);
-    console.log("tutup");
-  };
-
-  React.useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (!event.target.closest(".popup")) {
-        closePopup();
-      }
-    };
-
-    if (isPopupOpen) {
-      document.addEventListener("click", handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener("click", handleClickOutside);
-    };
-  }, [isPopupOpen]);
-
-  const handleDownload = (e) => {
-    e.stopPropagation();
-    e.preventDefault(); // Prevent the default behavior of the anchor element
-    const qrImage = 'assets/hehe.png'; // Path to the QR image
-    const link = document.createElement('a');
-    link.href = qrImage;
-    link.download = 'qr-image.png'; // Name of the downloaded file
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
-  
-
-  const handleCopyLink = () => {
-    const qrCodeLink = "https://shattereddisk.github.io/rickroll/rickroll.mp4";
-    navigator.clipboard.writeText(qrCodeLink)
-      .then(() => {
-        console.log("QR code link copied to clipboard:", qrCodeLink);
-        window.alert("QR code link copied to clipboard!");
-      })
-      .catch((error) => {
-        console.error("Failed to copy QR code link:", error);
-      });
-  };
-  
-
-  const handleIconClick = (e) => {
-    e.stopPropagation();
-    openPopup();
-  };
-  
-  return (
-    <div className="flex flex-col">
-      <img src={icon1} alt="" className="aspect-square w-[30px]" />
-      <img
-        src={icon2}
-        alt=""
-        className="self-center mt-36 aspect-square w-[38px] max-md:mt-10 cursor-pointer"
-        onClick={handleIconClick}
-      />
-      {isPopupOpen && (
-        <div className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50 z-50" onClick={closePopup}>
-          <div className="bg-white p-2 rounded-xl shadow-lg max-w-md" onClick={(e) => e.stopPropagation()}>
-            <img
-              src="assets/hehe.png"
-              alt="QR Code"
-              className="mx-auto mt-4"
-            />
-            <div className="flex justify-between -mt-1 relative">
-              <hr className="absolute -top-3 w-full border-t border-gray-300" />
-                <button onClick={handleDownload} className="text-white px-24 py-2 rounded-md">
-                  <img src="assets/DownloadIcon.png" alt="Download Icon" className="w-8 h-8" />
-                </button>
-              <hr className="absolute top-0 right-52 h-full border-l border-gray-300" />
-                <button onClick={handleCopyLink} className="text-white px-24 py-2 rounded-md">
-                  <img src="assets/CopyLinkIcon.png" alt="Copy Link Icon" className="w-8 h-8" />
-                </button>
-            </div>
-          </div>
-        </div>
-      )}
-    </div>
-  );  
-}
-
-
-
-function FeaturedEvent({ date, title, time }) {
-  const [day, month] = date.split(" ");
-
-  const handleMoreClick = () => { 
-    
-    console.log('More clicked!');
-    // Add more code here to perform specific actions, like opening a modal or navigating to a new page 
-    window.location.href = 'https://www.example.com/';
-  };
-
-  return (
-    <div className="flex gap-3 p-2 border-b border-l-gray-200">
-      <div className="flex flex-col pb-2.5 font-bold text-center whitespace-nowrap bg-red-500 rounded-lg shadow-sm h-[64px] w-[54px]">
-        <div className="justify-center px-5 py-3 text-xs bg-white rounded-t-md text-neutral-800">
-          {day}
-        </div>
-        <div className="self-center mt-1 text-xs text-white uppercase">
-          {month}
-        </div>
-      </div>
-      <div className="flex flex-col self-start text-sm font-semibold">
-        <div className="font-bold text-neutral-800">{title}</div>
-        <div className="mt-.5 text-neutral-400">{time}</div>
-        {/* Make the "More" text clickable */}
-        <div
-          className="mt-2.5 text-xs text-sky-500 underline cursor-pointer"
-          onClick={handleMoreClick} // Attach the click event handler 
-        > 
-          More
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function OnlinePerson({ name, imageUrl }) {
-  return (
-      <img
-        loading="lazy"
-        src={imageUrl}
-        alt={name}
-        className="shrink-0 w-11 aspect-[0.95]"
-      />
-  );
-}
-
-function Popup({ title, content, onClose, onSave }) {
-  return (
-    <div className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-50 z-50">
-      <div className="bg-white p-8 rounded-lg shadow-lg max-w-md">
-        <h2 className="text-2xl font-bold mb-4">{title}</h2>
-        <p className="mb-4">{content}</p>
-        <div className="flex justify-end">
-          <button onClick={onClose} className="mr-4 bg-gray-200 px-4 py-2 rounded-md">Cancel</button>
-          <button onClick={onSave} className="bg-blue-500 text-white px-4 py-2 rounded-md">Save</button>
-        </div>
-      </div>
-    </div>
-  );
-}
 
 export default function MyComponent() {
